requirements.txt
flask
flask-login
flask-oauthlib
flask-restful
flask-sqlalchemy
flask-whooshalchemy
mysql-python
uwsgi
